fix!(tests): better clean-up for ostormock
This commit uses t.Cleanup() to close all goroutines in ostormock
going forward. This requires the mock to be setup before each test
as reflected in the change in ostor_test.go.

Additionally, I added goleak to ensure I catch leaks going forward
so this should stabilize for consumers of the ostormock package.